fix an issue that can occur with IA3 if the PEFT params have a different dtype to the pretrained params

+++ src/peft/tuners/ia3/layer.py
@@ -92,7 +92,7 @@ class Linear(nn.Linear, IA3Layer):
             return
 
         self.weight = transpose(self.weight, self.fan_in_fan_out)
-        self.weight.data = torch.mul(self.weight.data, self.ia3_l[self.active_adapter].data)
+        self.weight.data = torch.mul(self.weight.data, self.ia3_l[self.active_adapter].data).to(self.weight.dtype)
         self.weight = transpose(self.weight, self.fan_in_fan_out)
 
         self.merged = True
